refactor: Move report-problem tool into the dev folder
The folder now matches its category (dev). Keep the report_problem.ts
filename and update the seven importers plus src/tools/AGENTS.md.
