+++ src/telemetry.ts
@@ -6,7 +6,7 @@ import type { z } from 'zod';
 import log from '@apify/log';
 
 import { DEFAULT_TELEMETRY_ENV, TELEMETRY_ENV } from './const.js';
-import type { reportProblemArgsSchema } from './tools/report_problem/report_problem.js';
+import type { reportProblemArgsSchema } from './tools/dev/report_problem.js';
 import type { ReportedProblemTelemetryProperties, TelemetryEnv, ToolCallTelemetryProperties } from './types.js';
 
 type ReportProblemArgs = z.infer<typeof reportProblemArgsSchema>;
