+++ src/tools/registry.ts
@@ -26,9 +26,9 @@ import { addActor } from './actors/add_actor.js';
 import { callActorApps, callActorDefault } from './actors/call_actor.js';
 import { fetchActorDetails } from './actors/fetch_actor_details.js';
 import { searchActors } from './actors/search_actors.js';
+import { reportProblem } from './dev/report_problem.js';
 import { fetchApifyDocs } from './docs/fetch_apify_docs.js';
 import { searchApifyDocs } from './docs/search_apify_docs.js';
-import { reportProblem } from './report_problem/report_problem.js';
 import { abortActorRun } from './runs/abort_actor_run.js';
 import { getActorRun } from './runs/get_actor_run.js';
 import { getActorRunList } from './runs/get_actor_run_list.js';
